debug, problema projected_slice_height más grande que screen height

+++ inc/cub3d.h
@@ -37,20 +37,20 @@
 
 typedef struct s_line_h
 {
-	double	intersection_x;
-	double	intersection_y;
-	double	xa;
-	double	ya;
+	int	intersection_x;
+	int	intersection_y;
+	int	xa;
+	int	ya;
 	double	perp_dist;
 	double	correct_dist;
 } 	t_line_h;
 
 typedef struct s_line_v
 {
-	double	intersection_x;
-	double	intersection_y;
-	double	xa;
-	double	ya;
+	int	intersection_x;
+	int	intersection_y;
+	int	xa;
+	int	ya;
 	double	perp_dist;
 	double	correct_dist;
 }	t_line_v;
